moved private bo data to test mock
